Handlers as module named web [Refactor: no logic change]
